requirements.txt
flask
pandas
sentence-transformers
faiss-cpu
groq
